Excise pyramid_swagger.swagger_version
